refactor(docs.ws): Merge deployed contracts components & imports
