+++ apps/docs.blocksense.network/components/DeployedContracts/DeployedContracts.tsx
@@ -1,117 +1,131 @@
-import * as React from 'react';
-import { Brand } from 'effect/Brand';
+'use client';
+
+import { useRef, useState } from 'react';
+import { Callout } from 'nextra/components';
+
 import { parseNetworkName } from '@blocksense/base-utils/evm';
-import CHAINLINK_COMPATIBILITY from '@blocksense/data-feeds-config-generator/chainlink_compatibility';
-import CONTRACTS_DEPLOYMENT_CONFIG from '@blocksense/data-feeds-config-generator/evm_contracts_deployment_v1';
-import {
-  decodeChainlinkCompatibilityConfig,
-  decodeDeploymentConfig,
-  DeploymentConfig,
-} from '@blocksense/config-types';
+
 import {
   CoreContract,
   ProxyContractData,
 } from '@/src/deployed-contracts/types';
-import { DeployedContractsClient } from '@/components/DeployedContracts/DeployedContractsClient';
+import { DataTable } from '@/components/ui/DataTable/DataTable';
+import {
+  columns as proxyContractsColumns,
+  proxyColumnsTitles,
+} from '@/components/DeployedContracts/proxyContractsColumns';
+import { ContractItemWrapper } from '@/components/sol-contracts/ContractItemWrapper';
+import { CoreContractCard } from '@/components/DeployedContracts/CoreContractCard';
+import { NetworkIcon } from '@/components/DeployedContracts/NetworkIcon';
+import { dataFeedUrl } from '@/src/constants';
 
-export function getContractsData() {
-  const networksData = decodeDeploymentConfig(CONTRACTS_DEPLOYMENT_CONFIG);
-
-  const getCoreContractsData = (networksData: DeploymentConfig) => {
-    const parsedCoreContracts: CoreContract[] = [];
-
-    Object.entries(networksData).forEach(([_networkName, networkData]) => {
-      if (!networkData) return;
-      if (_networkName === 'local') return;
-      const networkName = parseNetworkName(_networkName);
-      const coreContracts = networkData.contracts.coreContracts;
-
-      Object.entries(coreContracts).forEach(([contractName, contractsData]) => {
-        const existingContract = parsedCoreContracts.find(
-          contract => contract.contract === contractName,
-        );
-
-        if (existingContract) {
-          existingContract.networks.push(networkName);
-        } else {
-          parsedCoreContracts.push({
-            contract: contractName,
-            address: contractsData.address,
-            networks: [networkName],
-          });
-        }
-      });
-    });
-    return parsedCoreContracts;
-  };
-
-  const getProxyContractsContent = (networksData: DeploymentConfig) => {
-    type ChainLinkProxyData =
-      | (string &
-          Brand<'Hex String'> &
-          Brand<'Unformatted Data'> &
-          Brand<'EthereumAddress'>)
-      | undefined
-      | null;
-
-    const { blocksenseFeedsCompatibility } = decodeChainlinkCompatibilityConfig(
-      CHAINLINK_COMPATIBILITY,
-    );
-    const supportedNetworks: ProxyContractData[] = Object.entries(networksData)
-      .map(([_networkName, networkData]) => {
-        if (!networkData) return [];
-        if (_networkName === 'local') return [];
-        const networkName = parseNetworkName(_networkName);
-        const { CLAggregatorAdapter } = networkData.contracts;
-
-        return CLAggregatorAdapter.map(proxy => {
-          const compatibilityData = Object.entries(
-            blocksenseFeedsCompatibility,
-          ).find(([_id, data]) => data.description === proxy.description)?.[1];
-
-          if (!compatibilityData) {
-            throw new Error(
-              `No compatibility data found for ${proxy.description}`,
-            );
-          }
-
-          const chainLinkProxyData =
-            Object.entries(
-              compatibilityData.chainlink_compatibility.chainlink_aggregators,
-            ).find(([network, _data]) => network === networkName)?.[1] ?? null;
-
-          return {
-            ...proxy,
-            id: compatibilityData.id,
-            network: networkName,
-            chainlink_proxy: chainLinkProxyData,
-          };
-        });
-      })
-      .flat();
-    return supportedNetworks;
-  };
-
-  const parsedCoreContracts = getCoreContractsData(networksData);
-  const parsedProxyContracts = getProxyContractsContent(networksData);
-
-  return {
-    parsedCoreContracts,
-    parsedProxyContracts,
-  };
-}
-
-type FeedsDataProps = {
+type DeployedContractsProps = {
   parsedCoreContracts: CoreContract[];
   parsedProxyContracts: ProxyContractData[];
 };
 
 export const DeployedContracts = ({
-  parsedCoreContracts,
-  parsedProxyContracts,
-}: FeedsDataProps) => (
-  <DeployedContractsClient
-    parsedCoreContracts={parsedCoreContracts}
-    parsedProxyContracts={parsedProxyContracts}
-  />
-);
+  parsedCoreContracts: deployedCoreContracts,
+  parsedProxyContracts: deployedProxyContracts,
+}: DeployedContractsProps) => {
+  const [selectedNetwork, setSelectedNetwork] = useState<string | null>(null);
+  const contractsRef = useRef<HTMLDivElement | null>(null);
+
+  const handleNetworkClick = (network: string) => {
+    setSelectedNetwork(network);
+    setTimeout(() => {
+      contractsRef.current?.scrollIntoView({ behavior: 'smooth' });
+    }, 300);
+  };
+
+  return (
+    <section className="mt-4">
+      <ContractItemWrapper
+        title="Supported Networks"
+        titleLevel={2}
+        itemsLength={1}
+      >
+        <Callout type="info" emoji="💡">
+          We have deployed our contracts on the following networks. Select a
+          network to view detailed information about the deployed contracts.
+        </Callout>
+        <div className="flex flex-wrap justify-center gap-4 pt-4">
+          {deployedCoreContracts[0].networks.map(network => (
+            <NetworkIcon
+              key={network}
+              network={network}
+              isSelected={selectedNetwork === network}
+              onClick={() => {
+                handleNetworkClick(network);
+              }}
+            />
+          ))}
+        </div>
+      </ContractItemWrapper>
+      {selectedNetwork && (
+        <div ref={contractsRef}>
+          <ContractItemWrapper
+            title="Core Contracts"
+            titleLevel={2}
+            itemsLength={deployedCoreContracts.length}
+          >
+            <Callout type="info" emoji="💡">
+              <span>
+                These contracts are key components of the Blocksense platform
+                and provide essential functionalities that support the
+                ecosystem.
+                <br />
+                Discover more into our smart contracts
+                <a
+                  href={'./#smart-contract-architecture'}
+                  className="nx-text-primary-600 nx-underline nx-decoration-from-font [text-underline-position:from-font] mx-1"
+                >
+                  architecture
+                </a>
+                documentation section.
+              </span>
+            </Callout>
+            <div className="container px-0">
+              {deployedCoreContracts.map(contract => (
+                <CoreContractCard
+                  key={contract.address}
+                  contract={{
+                    name: contract.contract,
+                    address: contract.address,
+                    networks: contract.networks.filter(
+                      network => network === parseNetworkName(selectedNetwork),
+                    ),
+                  }}
+                />
+              ))}
+            </div>
+          </ContractItemWrapper>
+          <div className="mt-6">
+            <ContractItemWrapper
+              title="Aggregator Proxy Contracts"
+              titleLevel={2}
+              itemsLength={deployedProxyContracts.length}
+            >
+              <Callout type="info" emoji="💡">
+                Blocksense aggregator proxy contracts table allows users to
+                explore contracts that serve as an alternative to the Chainlink
+                proxy contracts. Additionally, the table provides information
+                about data feed names, IDs, and relevant addresses.
+              </Callout>
+              <DataTable
+                columns={proxyContractsColumns}
+                data={deployedProxyContracts.filter(
+                  element =>
+                    element.network === parseNetworkName(selectedNetwork),
+                )}
+                columnsTitles={proxyColumnsTitles}
+                filterCell="description"
+                rowLink={dataFeedUrl}
+              />
+            </ContractItemWrapper>
+          </div>
+        </div>
+      )}
+    </section>
+  );
+};
