--- apps/docs.blocksense.network/components/DeployedContracts/DeployedContractsClient.tsx
@@ -1,133 +1,0 @@
-'use client';
-import * as React from 'react';
-import { parseNetworkName } from '@blocksense/base-utils/evm';
-import {
-  CoreContract,
-  ProxyContractData,
-} from '@/src/deployed-contracts/types';
-import { DataTable } from '@/components/ui/DataTable/DataTable';
-import {
-  columns as proxyContractsColumns,
-  proxyColumnsTitles,
-} from '@/components/DeployedContracts/proxyContractsColumns';
-import { ContractItemWrapper } from '@/components/sol-contracts/ContractItemWrapper';
-import { CoreContractCard } from '@/components/DeployedContracts/CoreContractCard';
-import { dataFeedUrl } from '@/src/constants';
-import { NetworkIcon } from '@/components/DeployedContracts/NetworkIcon';
-import { Callout } from 'nextra/components';
-
-type FeedsDataProps = {
-  parsedCoreContracts: CoreContract[];
-  parsedProxyContracts: ProxyContractData[];
-};
-
-export const DeployedContractsClient = ({
-  parsedCoreContracts: deployedCoreContracts,
-  parsedProxyContracts: deployedProxyContracts,
-}: FeedsDataProps) => {
-  const [selectedNetwork, setSelectedNetwork] = React.useState<string | null>(
-    null,
-  );
-  const contractsRef = React.useRef<HTMLDivElement | null>(null);
-
-  const handleNetworkClick = (network: string) => {
-    setSelectedNetwork(network);
-    setTimeout(() => {
-      contractsRef.current?.scrollIntoView({ behavior: 'smooth' });
-    }, 300);
-  };
-
-  const smartContractsUrl = './#smart-contract-architecture';
-
-  return (
-    <section className="mt-4">
-      <ContractItemWrapper
-        title="Supported Networks"
-        titleLevel={2}
-        itemsLength={1}
-      >
-        <Callout type="info" emoji="💡">
-          We have deployed our contracts on the following networks. Select a
-          network to view detailed information about the deployed contracts.
-        </Callout>
-
-        <div className="flex flex-wrap justify-center gap-4 pt-4">
-          {deployedCoreContracts[0].networks.map(network => (
-            <NetworkIcon
-              key={network}
-              network={network}
-              isSelected={selectedNetwork === network}
-              onClick={() => {
-                handleNetworkClick(network);
-              }}
-            />
-          ))}
-        </div>
-      </ContractItemWrapper>
-      {selectedNetwork && (
-        <div ref={contractsRef}>
-          <ContractItemWrapper
-            title="Core Contracts"
-            titleLevel={2}
-            itemsLength={deployedCoreContracts.length}
-          >
-            <Callout type="info" emoji="💡">
-              <span>
-                These contracts are key components of the Blocksense platform
-                and provide essential functionalities that support the
-                ecosystem.
-                <br />
-                Discover more into our smart contracts
-                <a
-                  href={smartContractsUrl}
-                  className="nx-text-primary-600 nx-underline nx-decoration-from-font [text-underline-position:from-font] mx-1"
-                >
-                  architecture
-                </a>
-                documentation section.
-              </span>
-            </Callout>
-            <div className="container px-0">
-              {deployedCoreContracts.map(contract => (
-                <CoreContractCard
-                  key={contract.address}
-                  contract={{
-                    name: contract.contract,
-                    address: contract.address,
-                    networks: contract.networks.filter(
-                      network => network === parseNetworkName(selectedNetwork),
-                    ),
-                  }}
-                />
-              ))}
-            </div>
-          </ContractItemWrapper>
-          <div className="mt-6">
-            <ContractItemWrapper
-              title="Aggregator Proxy Contracts"
-              titleLevel={2}
-              itemsLength={deployedProxyContracts.length}
-            >
-              <Callout type="info" emoji="💡">
-                Blocksense aggregator proxy contracts table allows users to
-                explore contracts that serve as an alternative to the Chainlink
-                proxy contracts. Additionally, the table provides information
-                about data feed names, IDs, and relevant addresses.
-              </Callout>
-              <DataTable
-                columns={proxyContractsColumns}
-                data={deployedProxyContracts.filter(
-                  element =>
-                    element.network === parseNetworkName(selectedNetwork),
-                )}
-                columnsTitles={proxyColumnsTitles}
-                filterCell="description"
-                rowLink={dataFeedUrl}
-              />
-            </ContractItemWrapper>
-          </div>
-        </div>
-      )}
-    </section>
-  );
-};
