+++ apps/docs.blocksense.network/components/sol-contracts/ContractAddress.tsx
@@ -1,4 +1,5 @@
 'use client';
+
 import * as React from 'react';
 import Link from 'next/link';
 import { useRouter } from 'next/navigation';
