+++ apps/docs.blocksense.network/components/ui/DataTable/DataTable.tsx
@@ -1,4 +1,5 @@
 'use client';
+
 import * as React from 'react';
 import {
   ColumnDef,
